#1342 Prevent the possibility of unchecking the "Administrator" role on the customer edit page, if it is the only customer with this role

+++ src/Presentation/Nop.Web/Administration/Controllers/CustomerController.cs
@@ -740,6 +740,13 @@ namespace Nop.Admin.Controllers
             model.Address.PrepareCustomAddressAttributes(address, _addressAttributeService, _addressAttributeParser);
         }
 
+        [NonAction]
+        private bool SecondAdminAccountExists(Customer customer)
+        {
+            var customers = _customerService.GetAllCustomers(customerRoleIds: new[] {_customerService.GetCustomerRoleBySystemName(SystemCustomerRoleNames.Administrators).Id});
+
+            return customers.Count(p => p.Active && p.Id != customer.Id) > 0;
+        }
         #endregion
 
         #region Customers
@@ -1065,7 +1072,12 @@ namespace Nop.Admin.Controllers
                 {
                     customer.AdminComment = model.AdminComment;
                     customer.IsTaxExempt = model.IsTaxExempt;
-                    customer.Active = model.Active;
+                    
+                    if(!customer.IsAdmin() || model.Active || SecondAdminAccountExists(customer))
+                        customer.Active = model.Active;
+                    else
+                        ErrorNotification(_localizationService.GetResource("Admin.Customers.Customers.AdminAccountShouldExist"));
+
                     //email
                     if (!String.IsNullOrWhiteSpace(model.Email))
                     {
@@ -1203,8 +1215,14 @@ namespace Nop.Admin.Controllers
                         else
                         {
                             //remove role
-                            if (customer.CustomerRoles.Count(cr => cr.Id == customerRole.Id) > 0)
-                                customer.CustomerRoles.Remove(customerRole);
+                            if (customerRole.SystemName == SystemCustomerRoleNames.Administrators &&
+                                !SecondAdminAccountExists(customer))
+                            {
+                                ErrorNotification(_localizationService.GetResource("Admin.Customers.Customers.AdminAccountShouldExist"));
+                                continue;
+                            }
+
+                            customer.CustomerRoles.Remove(customerRole);
                         }
                     }
                     _customerService.UpdateCustomer(customer);
@@ -1256,7 +1274,7 @@ namespace Nop.Admin.Controllers
             PrepareCustomerModel(model, customer, true);
             return View(model);
         }
-        
+
         [HttpPost, ActionName("Edit")]
         [FormValueRequired("changepassword")]
         public ActionResult ChangePassword(CustomerModel model)
@@ -1353,6 +1371,12 @@ namespace Nop.Admin.Controllers
 
             try
             {
+                if (customer.IsAdmin() && !SecondAdminAccountExists(customer))
+                {
+                    ErrorNotification(_localizationService.GetResource("Admin.Customers.Customers.AdminAccountShouldExist"));
+                    return RedirectToAction("Edit", new { id = customer.Id });
+                }
+
                 _customerService.DeleteCustomer(customer);
 
                 //remove newsletter subscription (if exists)
